+++ src/CommunityToolkit.Maui/Converters/BaseConverterOneWay.shared.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Globalization;
+using CommunityToolkit.Maui.Extensions;
 using Microsoft.Maui.Controls;
 
 namespace CommunityToolkit.Maui.Converters;
