+++ src/CommunityToolkit.Maui/Converters/TextCaseConverter.shared.cs
@@ -7,6 +7,24 @@ using Microsoft.Maui.Controls;
 namespace CommunityToolkit.Maui.Converters;
 
 /// <summary>
+/// The text cases that can be used with <see cref="TextCaseConverter"/> to convert text to a specific case.
+/// </summary>
+public enum TextCaseType
+{
+	/// <summary>Should not be converted</summary>
+	None,
+
+	/// <summary>Convert to uppercase</summary>
+	Upper,
+
+	/// <summary>Convert to lowercase</summary>
+	Lower,
+
+	/// <summary>Converts the first letter to upper only</summary>
+	FirstUpperRestLower,
+}
+
+/// <summary>
 /// Converts text (string, char) to certain case as specified with <see cref="Type"/> or the parameter of the Convert method.
 /// </summary>
 [ContentProperty(nameof(Type))]
